Extract ConfigurationSessionTest into separate plug-in
The ConfigurationSessionTest introduces a many dependencies to the
org.eclipse.core.tests.harness plug-in, as it has to define dependencies
to everything that is part of the minimal bundle set for a
ConfigurationSessionTest at runtime. The list of dependencies will
become even larger when migration session tests to JUnit 5.

In order to reduce the dependencies of org.eclipse.core.tests.harness to
what is actually necessary for the plug-in's content, this change moves
the ConfigurationSessionTest into a separate plug-in, so that only that
plug-in has the according set of dependencies.
