+++ runtime/tests/org.eclipse.core.tests.harness.configurationsession/META-INF/MANIFEST.MF
@@ -1,0 +1,26 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: Eclipse Configuration Session Tests
+Bundle-SymbolicName: org.eclipse.core.tests.harness.configurationsession;singleton:=true
+Bundle-Version: 3.0.0.qualifier
+Automatic-Module-Name: org.eclipse.core.tests.harness.configurationsession
+Bundle-RequiredExecutionEnvironment: JavaSE-17
+Require-Bundle: org.junit,
+ org.eclipse.core.tests.harness,
+ org.eclipse.test.performance,
+ org.eclipse.core.runtime;bundle-version="[3.29.0,4.0.0)",
+ org.eclipse.jdt.junit.runtime,
+ org.eclipse.jdt.junit4.runtime,
+ org.eclipse.pde.junit.runtime
+Bundle-ActivationPolicy: lazy
+Import-Package: javax.xml.parsers,
+ net.bytebuddy,
+ org.apiguardian.api,
+ org.assertj.core.api,
+ org.junit.jupiter.api,
+ org.junit.platform.suite.api,
+ org.junit.platform.commons,
+ org.opentest4j,
+ org.w3c.dom,
+ org.xml.sax
+Export-Package: org.eclipse.core.tests.harness.configurationsession
