+++ runtime/tests/org.eclipse.core.tests.harness.configurationsession/build.properties
@@ -1,0 +1,16 @@
+###############################################################################
+# Copyright (c) 2024 Vector Informatik GmbH and others.
+#
+# This program and the accompanying materials
+# are made available under the terms of the Eclipse Public License 2.0
+# which accompanies this distribution, and is available at
+# https://www.eclipse.org/legal/epl-2.0/
+#
+# SPDX-License-Identifier: EPL-2.0
+###############################################################################
+source.. = src/
+output.. = bin/
+bin.includes = .,\
+               about.html,\
+               META-INF/
+src.includes = about.html
