+++ runtime/tests/org.eclipse.core.tests.harness/META-INF/MANIFEST.MF
@@ -8,20 +8,8 @@ Export-Package: org.eclipse.core.tests.harness;version="2.0",
  org.eclipse.core.tests.session;version="2.0"
 Require-Bundle: org.junit,
  org.eclipse.test.performance,
- org.eclipse.core.runtime;bundle-version="[3.29.0,4.0.0)",
- org.eclipse.jdt.junit.runtime,
- org.eclipse.jdt.junit4.runtime,
- org.eclipse.pde.junit.runtime
+ org.eclipse.core.runtime;bundle-version="[3.29.0,4.0.0)"
 Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: JavaSE-17
-Import-Package: javax.xml.parsers,
- net.bytebuddy,
- org.apiguardian.api,
- org.assertj.core.api,
- org.junit.jupiter.api,
- org.junit.platform.suite.api,
- org.junit.platform.commons,
- org.opentest4j,
- org.w3c.dom,
- org.xml.sax
+Import-Package: org.assertj.core.api
 Automatic-Module-Name: org.eclipse.core.tests.harness
