+++ runtime/tests/org.eclipse.core.tests.runtime/META-INF/MANIFEST.MF
@@ -13,7 +13,8 @@ Export-Package: org.eclipse.core.tests.internal.preferences,
 Require-Bundle: org.junit,
  org.eclipse.test.performance;resolution:=optional,
  org.eclipse.core.runtime;bundle-version="3.29.0",
- org.eclipse.core.tests.harness;bundle-version="3.11.0"
+ org.eclipse.core.tests.harness;bundle-version="3.11.0",
+ org.eclipse.core.tests.harness.configurationsession
 Import-Package: org.assertj.core.api
 Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: JavaSE-17
